+++ packages/sdk/src/models/models.js
@@ -15,21 +15,6 @@ import * as recordColoring from './record_coloring';
 const permissionHelpers = window.__requirePrivateModuleFromAirtable(
     'client_server_shared/permissions/permission_helpers',
 );
-const hyperIdGenerator = window.__requirePrivateModuleFromAirtable(
-    'client_server_shared/hyper_id/hyper_id_generator',
-);
-
-/**
- * Helper to generate a GUID.
- *
- * @function
- * @returns {string}
- * @alias generateGuid
- * @example
- * import {generateGuid} from '@airtable/blocks/models';
- * const id = generateGuid();
- */
-const generateGuid = hyperIdGenerator.generateGuid;
 
 const models = {
     Base,
@@ -45,8 +30,6 @@ const models = {
     fieldTypes: FieldTypes,
     viewTypes: ViewTypes,
     permissionLevels: permissionHelpers.ApiPermissionLevels,
-
-    generateGuid,
 };
 
 export default models;