+++ packages/sdk/src/models/record_query_result.js
@@ -3,14 +3,13 @@ import Colors, {type Color} from '../colors';
 import {type BaseData} from '../types/base';
 import {type RecordId} from '../types/record';
 import {FieldTypes} from '../types/field';
-import {isEnumValue, assertEnumValue} from '../private_utils';
+import {isEnumValue, assertEnumValue, getLocallyUniqueId} from '../private_utils';
 import {
     spawnAbstractMethodError,
     spawnUnknownSwitchCaseError,
     spawnError,
     invariant,
 } from '../error_utils';
-import getSdk from '../get_sdk';
 import AbstractModelWithAsyncData from './abstract_model_with_async_data';
 import type Table from './table';
 import Field from './field';
@@ -330,7 +329,7 @@ class RecordQueryResult<DataType = {}> extends AbstractModelWithAsyncData<
      * @private
      */
     constructor(normalizedOpts: NormalizedRecordQueryResultOpts, baseData: BaseData) {
-        super(baseData, getSdk().models.generateGuid());
+        super(baseData, getLocallyUniqueId('RecordQueryResult'));
         this._normalizedOpts = normalizedOpts;
         this._recordStore = normalizedOpts.recordStore;
     }