+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 -   Upgrade flow to 0.106.3
 -   Fix a flow error with `react-window`.
+-   **BREAKING:** Remove `models.generateGuid()`. Use an ID generator like
+    [`uuid`](https://www.npmjs.com/package/uuid) instead.
 
 ## [0.0.30](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.29...@airtable/blocks@0.0.30) - 2019-08-26